Move pipeline reducers to their own file
