+++ lib/micro/service.rb
@@ -3,10 +3,10 @@
 require 'micro/attributes'
 
 require 'micro/service/version'
-
 require 'micro/service/error'
 require 'micro/service/result'
 require 'micro/service/base'
 require 'micro/service/safe'
 require 'micro/service/strict'
+require 'micro/service/pipeline/reducer'
 require 'micro/service/pipeline'
